refactor: extract OpenAI auth refresh recovery
