+++ .agents/plugins/opencode-aidevops/openai-provider-auth.mjs
@@ -6,16 +6,16 @@
  * usage-limited accounts and retries once with another pooled account.
  */
 
-import { getAccounts, markAuthRefreshFailure, patchAccount, rotateOpenAIPoolToken } from "./oauth-pool.mjs";
-import { OPENAI_TOKEN_ENDPOINT } from "./oauth-pool-constants.mjs";
+import { getAccounts, patchAccount, rotateOpenAIPoolToken } from "./oauth-pool.mjs";
 import { annotateOpenAIOverloadResponse, formatRetryDelay, isOpenAIOverloadText, overloadRetryDelaysMs, sleep, wrapOpenAIOverloadStream } from "./openai-overload-retry.mjs";
+import { handleOpenAITokenRefreshFailure, isOpenAITokenRefreshRequest, readOpenAITokenRefreshBody } from "./openai-auth-refresh-recovery.mjs";
 
 export { isOpenAIOverloadText } from "./openai-overload-retry.mjs";
+export { isOpenAITokenRefreshRequest } from "./openai-auth-refresh-recovery.mjs";
 
 const OPENAI_API_HOST = "api.openai.com";
 const OPENAI_API_PREFIX = "/v1/";
 const DEFAULT_OPENAI_COOLDOWN_MS = 300_000;
-const TOKEN_REFRESH_FAILURE_STATUSES = new Set([401, 403]);
 const USAGE_LIMIT_MARKERS = [
   "usage limit",
   "rate limit",
@@ -47,14 +47,6 @@ export function isOpenAIProviderRequest(input) {
   try {
     const url = new URL(requestUrl(input));
     return url.hostname === OPENAI_API_HOST && url.pathname.startsWith(OPENAI_API_PREFIX);
-  } catch { return false; }
-}
-
-export function isOpenAITokenRefreshRequest(input) {
-  try {
-    const url = new URL(requestUrl(input));
-    const endpoint = new URL(OPENAI_TOKEN_ENDPOINT);
-    return url.origin === endpoint.origin && url.pathname === endpoint.pathname;
   } catch { return false; }
 }
 
@@ -124,54 +116,6 @@ function buildRetryInit(input, init, accessToken) {
   return retryInit;
 }
 
-async function readRequestBodyText(input, init) {
-  try {
-    if (typeof init?.body === "string") return init.body;
-    if (init?.body instanceof URLSearchParams) return init.body.toString();
-    if (typeof Request !== "undefined" && input instanceof Request) return await input.clone().text();
-  } catch { /* best-effort: request body may be a non-cloneable stream */ }
-  return "";
-}
-
-function extractOpenAIRefreshToken(bodyText) {
-  try {
-    const params = new URLSearchParams(bodyText);
-    if (params.get("grant_type") !== "refresh_token") return "";
-    return params.get("refresh_token") || "";
-  } catch { return ""; }
-}
-
-function resolveOpenAIRefreshAccount(refreshToken) {
-  if (!refreshToken) return null;
-  return getAccounts("openai").find((account) => account.refresh === refreshToken) || null;
-}
-
-function buildOpenAITokenRefreshResponse(account) {
-  const expiresAt = Number(account.expires) || Date.now() + 3600_000;
-  const expiresIn = Math.max(1, Math.floor((expiresAt - Date.now()) / 1000));
-  return new Response(JSON.stringify({
-    access_token: account.access,
-    refresh_token: account.refresh,
-    expires_in: expiresIn,
-    token_type: "Bearer",
-  }), {
-    status: 200,
-    headers: { "content-type": "application/json" },
-  });
-}
-
-async function handleOpenAITokenRefreshFailure(ctx) {
-  const { client, response, refreshToken } = ctx;
-  if (!TOKEN_REFRESH_FAILURE_STATUSES.has(response.status)) return response;
-  const current = resolveOpenAIRefreshAccount(refreshToken);
-  const currentEmail = current?.email || "unknown";
-  if (current) markAuthRefreshFailure("openai", current);
-  console.error(`[aidevops] OpenAI provider: token refresh failed (${response.status}) for ${currentEmail} — rotating pool account`);
-  const rotated = await rotateOpenAIPoolToken(client, current?.email);
-  if (!rotated?.access) return response;
-  return buildOpenAITokenRefreshResponse(rotated);
-}
-
 async function handleOpenAIUsageLimit(ctx) {
   const { client, originalFetch, input, init, response, retryInput } = ctx;
   const current = resolveOpenAIAccount(extractBearerToken(input, init));
@@ -238,7 +182,7 @@ async function handleOpenAIFetchRequest(ctx) {
   const { client, originalFetch, input, init } = ctx;
   const openaiRequest = isOpenAIProviderRequest(input);
   const tokenRefreshRequest = isOpenAITokenRefreshRequest(input);
-  const tokenRefreshBody = tokenRefreshRequest ? await readRequestBodyText(input, init) : "";
+  const tokenRefreshBody = tokenRefreshRequest ? await readOpenAITokenRefreshBody(input, init) : "";
   const retryInput = openaiRequest ? buildRetryRequest(input) : input;
   const firstInit = openaiRequest ? await maybeRotateBeforeOpenAIFetch(client, input, init) : init;
   const response = await originalFetch(input, firstInit);
@@ -247,7 +191,7 @@ async function handleOpenAIFetchRequest(ctx) {
     return handleOpenAITokenRefreshFailure({
       client,
       response,
-      refreshToken: extractOpenAIRefreshToken(tokenRefreshBody),
+      bodyText: tokenRefreshBody,
     });
   }
   if (!openaiRequest) return response;
